N10-008 Subnetting Quiz Solutions
There are many ways of teaching and learning subnetting.  I have sat through them all and in the end, I find that there is no one common solution or method that covers all angles of application.  To be able to subnet, you must understand the basics and then enter each problem with a solution goal in mind.  For example, if your goal is determining the last usable IP address on a network, then don't waste time on determining the number of avail subnets.  In these solutions, I focused quickly on the question being asked to ensure the solutions were efficient.